requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
python-dotenv==1.0.0
supabase==2.0.0
google-generativeai==0.3.0
requests==2.31.0
# Optional: uncomment if you want to use ElevenLabs SDK
elevenlabs==1.0.0